pyproject.toml
[tool.poetry]
name = "pixellab"
version = "1.0.7"
description = ""
authors = ["PixelLab"]
readme = "README.md"
packages = [
    { include = "pixellab" }
]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.0.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
pillow = "*"
eval-type-backport = { version = "*", python = "<3.10" }

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
pytest = "^8.0.0"
ipykernel = "^6.29.0"
ipywidgets = "^8.1.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
